Removed non-essetial logging functions
