+++ core/utils.py
@@ -24,7 +24,6 @@ import torch.nn.functional as F
 import torchvision
 import torchvision.utils as vutils
 
-import logging
 
 
 def save_json(json_file, filename):
@@ -38,7 +37,6 @@ def print_network(network, name):
         num_params += p.numel()
     # print(network)
     print("Number of parameters of %s: %i" % (name, num_params))
-    logging.debug("Number of parameters of %s: %i" % (name, num_params))
 
 
 def he_init(module):
